+++ bapsflib/_hdf/utils/hdfreadmsi.py
