+++ docs/src/bapsflib._hdf.utils.hdfreadmsi.rst
@@ -1,0 +1,11 @@
+bapsflib\.\_hdf\.utils\.hdfreadmsi
+==================================
+
+.. automodule:: bapsflib._hdf.utils.hdfreadmsi
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+    .. rubric:: Classes
+
+    .. autosummary:: HDFReadMSI
