+++ docs/src/bapsflib._hdf.utils.rst
@@ -1,0 +1,11 @@
+bapsflib\.\_hdf\.utils
+======================
+
+.. automodule:: bapsflib._hdf.utils
+
+.. toctree::
+    :maxdepth: 2
+    :titlesonly:
+    :caption: Sub-Packages & Modules
+
+    bapsflib._hdf.utils.hdfreadmsi
