--- docs/src/bapsflib.lapd._hdf.hdfreadmsi.rst
@@ -1,11 +1,0 @@
-bapsflib\.lapd\.\_hdf\.hdfreadmsi
-=================================
-
-.. automodule:: bapsflib.lapd._hdf.hdfreadmsi
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-    .. rubric:: Classes
-
-    .. autosummary:: HDFReadMSI
